Finishing addressing Matthews comments
